refactor: move pages toward repository-driven rendering
