+++ src/collab/remoteEvents.js
@@ -1,0 +1,1 @@
+export const COLLAB_REMOTE_EVENT_ACTION = "collab.remote.event";
