+++ src/deps/features/resourcePages/media/mediaPageShared.js
@@ -1,16 +1,26 @@
 const EMPTY_TREE = { tree: [], items: {} };
 
-export const getMediaPageData = ({ projectService, resourceType } = {}) => {
-  return projectService.getState()[resourceType] ?? EMPTY_TREE;
+export const getMediaPageData = ({
+  repositoryState,
+  projectService,
+  resourceType,
+} = {}) => {
+  const state = repositoryState ?? projectService?.getState?.();
+  return state?.[resourceType] ?? EMPTY_TREE;
 };
 
 export const syncMediaPageData = ({
   store,
+  repositoryState,
   projectService,
   resourceType,
 } = {}) => {
   store.setItems({
-    data: getMediaPageData({ projectService, resourceType }),
+    data: getMediaPageData({
+      repositoryState,
+      projectService,
+      resourceType,
+    }),
   });
 };
