+++ src/deps/services/web/collab/constants.js
@@ -1,0 +1,4 @@
+export const DEFAULT_COLLAB_ENDPOINT = "wss://127.0.0.1:8787/sync";
+export const COLLAB_HEARTBEAT_INTERVAL_MS = 10_000;
+export const COLLAB_RECONNECT_INTERVAL_MS = 5_000;
+export const COLLAB_CONNECTION_ERROR_THROTTLE_MS = 10_000;
