+++ src/pages/fonts/fonts.handlers.js
@@ -67,6 +67,8 @@ const createFontsFromFiles = async ({ deps, files, parentId } = {}) => {
 };
 
 const {
+  handleBeforeMount,
+  handleAfterMount,
   refreshData: handleDataChanged,
   handleFileExplorerSelectionChanged,
   handleFileExplorerAction,
@@ -80,13 +82,9 @@ const {
 
 export { handleFileExplorerAction, handleFileExplorerTargetChanged };
 
-export const handleAfterMount = async (deps) => {
-  const { projectService } = deps;
-  await projectService.ensureRepository();
-  await handleDataChanged(deps);
-};
-
 export {
+  handleBeforeMount,
+  handleAfterMount,
   handleDataChanged,
   handleFileExplorerSelectionChanged,
   handleSearchInput,
