+++ src/pages/images/images.handlers.js
@@ -1,13 +1,11 @@
 import { nanoid } from "nanoid";
-import { filter, tap } from "rxjs";
 import { createMediaPageHandlers } from "../../deps/features/resourcePages/media/createMediaPageHandlers.js";
 import { resolveResourceParentId } from "../../deps/features/resourcePages/media/mediaPageShared.js";
-
-const COLLAB_IMAGES_REFRESH_ACTION = "collab.images.refresh";
 
 const {
   refreshData: handleDataChanged,
   handleBeforeMount,
+  handleAfterMount,
   handleFileExplorerSelectionChanged,
   handleFileExplorerDoubleClick,
   handleFileExplorerAction,
@@ -18,23 +16,13 @@ const {
   handleItemEdit: handleImageItemEdit,
 } = createMediaPageHandlers({
   resourceType: "images",
-  subscriptions: (deps) => {
-    const { subject } = deps;
-    return [
-      subject.pipe(
-        filter(({ action }) => action === COLLAB_IMAGES_REFRESH_ACTION),
-        tap(() => {
-          deps.handlers.handleDataChanged(deps);
-        }),
-      ),
-    ];
-  },
   selectItemById: (store, { itemId }) => store.selectImageItemById({ itemId }),
   getEditPreviewFileId: (item) => item?.fileId,
 });
 
 export {
   handleBeforeMount,
+  handleAfterMount,
   handleFileExplorerSelectionChanged,
   handleFileExplorerDoubleClick,
   handleFileExplorerAction,
