+++ src/pages/transforms/transforms.handlers.js
@@ -133,6 +133,7 @@ const openTransformDialog = async ({
 };
 
 const {
+  handleBeforeMount,
   handleAfterMount,
   refreshData: handleDataChanged,
   handleFileExplorerSelectionChanged,
@@ -145,6 +146,7 @@ const {
 });
 
 export {
+  handleBeforeMount,
   handleAfterMount,
   handleDataChanged,
   handleFileExplorerSelectionChanged,
