+++ src/pages/videos/videos.handlers.js
@@ -95,6 +95,7 @@ const openVideoPreviewById = async ({ deps, itemId } = {}) => {
 
 const {
   handleBeforeMount,
+  handleAfterMount,
   refreshData: handleDataChanged,
   handleFileExplorerSelectionChanged,
   handleFileExplorerDoubleClick,
@@ -112,6 +113,7 @@ const {
 
 export {
   handleBeforeMount,
+  handleAfterMount,
   handleDataChanged,
   handleFileExplorerSelectionChanged,
   handleFileExplorerDoubleClick,
